Refactor project into smaller components
