+++ src/lib/components/calendar/overlay.js
